+++ fiddles/aurelia/template/app.js
@@ -1,3 +1,0 @@
-export class App {
-  message = 'hello world';
-}